example/requirements.txt
transformers==4.46.3
accelerate==0.34.2
datasets==3.1.0
torch==2.5.1
evaluate==0.4.3
scikit-learn==1.5.2
PyYAML==6.0.2
numpy==2.1.3
pytest==8.3.3
jsonschema==4.21.1
